moved correct_cast to a casting module, to prevent circular imports
