+++ scripts/dynamic_programming.py
@@ -9,10 +9,11 @@ import torch
 from configs import EnvConfig
 from simple_parsing import ArgumentParser
 
+from gfn.casting import correct_cast
 from gfn.estimators import LogEdgeFlowEstimator
 from gfn.losses import FMParametrization
 from gfn.modules import Tabular, Uniform
-from gfn.utils import validate, correct_cast
+from gfn.utils import validate
 
 parser = ArgumentParser()
