+++ src/gfn/samplers/actions_samplers.py
@@ -5,9 +5,9 @@ import torch
 from torch.distributions import Categorical
 from torchtyping import TensorType
 
+from gfn.casting import correct_cast
 from gfn.containers import States
 from gfn.estimators import LogEdgeFlowEstimator, LogitPBEstimator, LogitPFEstimator
-from gfn.utils import correct_cast
 
 # Typing
 Tensor2D = TensorType["batch_size", "n_actions"]
