+++ src/gfn/utils.py
@@ -1,7 +1,6 @@
-from typing import Dict, Optional, cast
+from typing import Dict, Optional
 
 import torch
-from torchtyping import TensorType
 
 from gfn.containers import States, Trajectories, Transitions
 from gfn.distributions import EmpiricalTerminatingStatesDistribution
@@ -14,23 +13,6 @@ from gfn.losses import (
     TBParametrization,
     TrajectoryDecomposableLoss,
 )
-
-# Typing.
-ForwardMasksTensor = TensorType["batch_shape", "n_actions", torch.bool]
-BackwardMasksTensor = TensorType["batch_shape", "n_actions - 1", torch.bool]
-
-
-def correct_cast(
-    forward_masks: ForwardMasksTensor | None,
-    backward_masks: BackwardMasksTensor | None,
-) -> tuple[ForwardMasksTensor, BackwardMasksTensor]:
-    """
-    Casts the given masks to the correct type, if they are not None.
-    This function is to help with type checking only.
-    """
-    forward_masks = cast(ForwardMasksTensor, forward_masks)
-    backward_masks = cast(BackwardMasksTensor, backward_masks)
-    return forward_masks, backward_masks
 
 
 def trajectories_to_training_samples(
